// src/app/models/master/masConsentDiscloseData.ts
export interface MasConsentDiscloseData {

    consentDiscloseId?: number;
    consentId?: number;
    consentDiscloseTh?: string;
    consentDiscloseEn?: string;
    createDate?: Date;
    createById?: number;
    updateDate?: Date;
    updateById?: number;
    activeFlag?: boolean;
    rowNum?:number;
    rows?:number;
    first?: number;
    size?: number;

}